requirements.txt
numpy
pandas
sklearn
matplotlib
python-dotenv
mysql-connector
sqlalchemy